RemoteS

Index

RemoteS.bright_T

RemoteS.clg

RemoteS.clre

RemoteS.dn2radiance

RemoteS.dn2reflectance

RemoteS.dn2temperature

RemoteS.evi

RemoteS.evi2

RemoteS.gndvi

RemoteS.grid_at_sensor

RemoteS.mcari

RemoteS.mndwi

RemoteS.msavi

RemoteS.mtci

RemoteS.nbri

RemoteS.ndrei1

RemoteS.ndrei2

RemoteS.ndvi

RemoteS.ndvi

RemoteS.ndwi

RemoteS.ndwi2

RemoteS.radiance_TOA

RemoteS.read_mtl

RemoteS.reflectance_TOA

RemoteS.reflectance_surf

RemoteS.satvi

RemoteS.savi

RemoteS.slavi

RemoteS.spectral_indices

RemoteS.truecolor

Functions

RemoteS.bright_T — Method

R = bright_T(fname::String)

Returns a GMTgrid with the brigthness temperature of Landasat8 termal band (10 or 11)

RemoteS.clg — Method

CLG = clg(green, redEdge3; kw...)

Green cholorphyl index. Wu et al 2012.

CLG = (redEdge3)/(green)-1

RemoteS.clre — Method

CLRE = clre(redEdge1, redEdge3; kw...)

RedEdge cholorphyl index. Clevers and Gitelson 2013.

CLRE = (redEdge3)/(redEdge1)-1

RemoteS.dn2radiance — Method

R = dn2radiance(fname::String)

Returns a GMTgrid with the radiance at TopOfAtmosphere for the Landsat8 band file fname

RemoteS.dn2reflectance — Method

R = dn2reflectance(fname::String)

Returns a GMTgrid with the TopOfAtmosphere planetary reflectance for the Landsat8 band file fname

RemoteS.dn2temperature — Method

R = dn2temperature(fname::String)

Returns a GMTgrid with the brigthness temperature of Landasat8 termal band (10 or 11)

RemoteS.evi — Method

EVI = evi(blue, red, nir; kw...)

Enhanced vegetation index. Huete et al 1990

EVI = G * ((nir - red) / (nir + C1 * red - C2 * blue + Levi)); C1, C2, G, Levi = 6.0, 7.5, 2.5, 1.

RemoteS.evi2 — Method

EVI2 = evi(red, nir; kw...)

Two-band Enhanced vegetation index. Jiang et al 2008

EVI2 = G * ((nir - red) / (nir + 2.4 * red ))

RemoteS.gndvi — Method

GNDVI = gndvi(green, nir; kw...)

green Normalized diff vegetation index: more sensitive to cholorphyll than ndvi. Gitelson, A., and M. Merzlyak

GNDVI = (nir - green) / (nir + green)

RemoteS.grid_at_sensor — Function

G = grid_at_sensor(fname::String, sds_name::String=""; V::Bool=false, kw...)

Read one of those netCDF files that are not regular grids but have instead the coordinates in the LONGITUDE abd LATITUDE arrays. MODIS L2 files are a good example of this. Data in theses files are not layed down on a regular grid and we must interpolate to get one. Normally the lon and lat arrays are called 'longitude' and 'latitude' and these it's what is seek for by default. But files exist that pretend to comply to CF but use other names. In this case, use the kwargs 'xarray' & 'yarray' to pass in the variable names. For example: xarray="XLONG", yarray="XLAT" The other fundamental info to pass in is the name of the array to be read/interpolated. We do that via the SDS_NAME arg.

In simpler cases the variable to be interpolated lays down on a 2D array but it is also possible that it is stored in a 3D array. If that is the case, use the keyword 'band' to select a band (ex: 'band=2') Bands are numbered from 1.

The interpolation is done so far with 'nearneighbor'. Both the region (-R) and increment (-I) are estimated from data but they can be set with 'region' and 'inc' kwargs as well. For MODIS data we can select the quality flag to filter by data quality. By default the best quality (=0) is used, but one can select another with the quality=val kwarg. Positive 'val' values select data of quality <= quality, whilst negative 'val' values select only data with quality >= abs(val). This allows for example to extract only the cloud coverage.

If instead of calculating a grid (returned as a GMTgrid type) user wants the x,y,z data intself, use the keywords 'dataset', or 'outxyz' and the output will be in a GMTdataset (i.e. use 'dataset=true').

To inquire just the list of available arrays use 'list=true' or 'gdalinfo=true' to get the full file info.

Examples:

G = grid_at_sensor("AQUA_MODIS.20020717T135006.L2.SST.nc", "sst", V=true);

G = grid_at_sensor("TXx-narr-annual-timavg.nc", "T2MAX", xarray="XLONG", yarray="XLAT", V=true);

RemoteS.mcari — Method

MCARI = mcari(green, red, redEdge1; kw...)

Modified Chlorophyll Absorption ratio index. Daughtery et al. 2000

MCARI = (redEdge1 - red - 0.2 * (redEdge1 + green)) * (redEdge1 / red)

RemoteS.mndwi — Method

MNDWI = mndwi(green, swir2; kw...)

Modified Normalised Difference Water Index. Xu2006

MNDWI = (green-swir2) / (green+swir2)

RemoteS.msavi — Method

MSAVI = msavi(red, nir; kw...)

Modified soil adjusted vegetation index. Qi 1994

MSAVI = nir + 0.5 - (0.5 * sqrt(pow(2.0 * nir + 1.0, 2) - 8.0 * (nir - (2.0 * red))))

RemoteS.mtci — Method

MTCI = mtci(red, redEdge1, redEdge2; kw...)

Meris Terrestrial Chlorophyll Index. Clevers and Gitelson 2013, Dash and Curran 2004

MTCI = (redEdge2-redEdge1) / (redEdge1-red)

RemoteS.nbri — Method

NBRI = nbri(nir, swir3; kw...)

Normalised Burn Ratio Index. Garcia 1991

NBRI = (nir - swir3) / (nir + swir3)

RemoteS.ndrei1 — Method

NDREI1 = ndrei1(redEdge1, redEdge2; kw...)

Normalized difference red edge index. Gitelson and Merzlyak 1994

NDREI1 = (redEdge2 - redEdge1) / (redEdge2 + redEdge1)

RemoteS.ndrei2 — Method

NDREI2 = ndrei2(redEdge1, redEdge3; kw...)

Normalized difference red edge index 2. Barnes et al 2000

NDREI2 = (redEdge3 - redEdge1) / (redEdge3 + redEdge1)

RemoteS.ndvi — Method

NDVI = ndvi(red, nir; kw...)

Compute the NDVI vegetation index. Input can be either the bands file names, or GMTimage objects with the band's data.

NDVI = (nir - red) / (nir + red)

Returns either a Float32 GMTgrid or a UInt8 GMTimage if the mask option is set to true.

RemoteS.ndvi — Method

NDVI = ndvi(bndR, bndNIR; threshold=0.4, mask::Bool=false)

Compute the NDVI vegetation index. Input can be either the bands file names, or GMTimage objects with the band's data.

Returns either a Float32 GMTgrid or a UInt8 GMTimage if the mask option is set to true.

RemoteS.ndwi — Method

NDWI = ndwi(green, nir; kw...)

Normalized difference water index. McFeeters 1996. NDWI => (green - nir)/(green + nir)

NDWI = (green - nir)/(green + nir)

RemoteS.ndwi2 — Method

NDWI2 = ndwi2(nir, swir2; kw...)

Normalized difference water index. Gao 1996, Chen 2005 (also known as Normalized Difference Moisture Index NDBI and LSWI)

NDWI2 = (nir - swir2)/(nir + swir2)

RemoteS.radiance_TOA — Method

R = radiance_TOA(fname::String)

Returns a GMTgrid with the radiance at TopOfAtmosphere for the Landsat8 band file fname

RemoteS.read_mtl — Function

readmtl(bandname::String, mtl::String="")

Use the band_name of a Landsat8 band to find the MTL file with the parameters of the scene at which band belongs and read the params needed to compute Brightness temperature, radiance at top of atmosphere, etc. If the MTL file does not lieves next to the band file, send its name via the mtl argument.

Returns a tuple with:

(band=band, radmul=radmul, radadd=radadd, radmax=radmax, reflectmul=reflectmul, reflectadd=reflectadd, reflectmax=reflectmax, sunazim=sunazim, sunelev=sunelev, sundis=sunazim, K1=K1, K2=K2)

RemoteS.reflectance_TOA — Method

R = reflectance_TOA(fname::String)

Returns a GMTgrid with the TopOfAtmosphere planetary reflectance for the Landsat8 band file fname

RemoteS.reflectance_surf — Method

R = reflectance_surf(fname::String)

Compute the radiance-at-surface of Landsat8 band using the COST model.

Returns a Float32 GMTgrid type

RemoteS.satvi — Method

SATVI = satvi(red, swir2, swir3; kw...)

Soil adjusted total vegetation index. Marsett 2006

SATVI = ((swir2 - red) / (swir2 + red + L)) * (1.0 + L) - (swir3 / 2.0)

RemoteS.savi — Method

SAVI = savi(red, nir; kw...)

Soil adjusted vegetation index. Huete 1988

SAVI = (nir - red) * (1.0 + L) / (nir + red + L)

RemoteS.slavi — Method

SLAVI = slavi(red, nir, swir2; kw...)

Specific Leaf Area Vegetation Index. Lymburger 2000

SLAVI = nir / (red + swir2)

RemoteS.spectral_indices — Function

RemoteS.truecolor — Method

Irgb = truecolor(bndR, bndG, bndB)

Take three Landsat8/Sentinel2 UINT16 GMTimages or the file names of those bands and compose an RGB true color image applying automatic histogram stretching.

Return an UInt8 RGB GMTimage